Fix issues when testing job
